+++ src/main/java/com/jiyong/commerce/common/util/CommonUtils.java
@@ -3,7 +3,7 @@ package com.jiyong.commerce.common.util;
 
 import java.util.concurrent.ThreadLocalRandom;
 
-public class AspectUtils {
+public class CommonUtils {
 
     public static void sleep(int delay) {
         try {
@@ -14,6 +14,9 @@ public class AspectUtils {
     }
 
     public static int getRandomNumber(int min, int max) {
+        if (min > max) {
+            throw new IllegalArgumentException("min cannot be greater than max!!");
+        }
         return ThreadLocalRandom.current().nextInt(min, max + 1);
     }
